+++ crates/swc_estree_ast/src/comment.rs
@@ -1,0 +1,27 @@
+
+
+    #[derive(Debug, Clone, PartialEq)]
+    #[ast_serde("BaseComment")]
+    pub struct BaseComment {
+        pub value: String,
+        pub start: usize,
+        pub end: usize,
+        pub loc: Loc,
+    }
+
+    #[derive(Debug, Clone, PartialEq)]
+    #[ast_serde]
+    pub enum Comment {
+        #[tag("CommentBlock")]
+        Block(BaseComment),
+        #[tag("CommentLine")]
+        Line(BaseComment),
+    }
+
+    #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+    #[serde(rename_all = "lowercase")]
+    pub enum CommentTypeShorthand {
+        Leading,
+        Inner,
+        Trailing,
+    }
